+++ crates/tests/src/system_tests/mod.rs
@@ -1,5 +1,2 @@
-pub mod bad_block;
-pub mod bad_challenge;
-pub mod deposit;
 pub mod test_mode_control;
 pub mod utils;
